# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "بريمو - Bremo — API v1",
        "version": "1.0.0",
        "description": "مولَّد من مسارات لارافيل. المرجع الملزم: docs/31-API-CONTRACT. لا يوجد أي مسار لتعديل حالة الطلب مباشرة؛ كل تغيير يمر بانتقال في docs/10."
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "parameters": {
            "AppMode": {
                "name": "X-App-Mode",
                "in": "header",
                "required": false,
                "schema": {
                    "type": "string",
                    "enum": [
                        "CUSTOMER",
                        "PROVIDER"
                    ]
                },
                "description": "يحدد ما يُعرض فقط؛ الصلاحية من الخادم (23)."
            },
            "IdempotencyKey": {
                "name": "Idempotency-Key",
                "in": "header",
                "required": false,
                "schema": {
                    "type": "string",
                    "format": "uuid"
                },
                "description": "نفس المفتاح خلال 24 ساعة يعيد نفس النتيجة."
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "object",
                        "required": [
                            "code",
                            "message"
                        ],
                        "properties": {
                            "code": {
                                "type": "string",
                                "enum": [
                                    "VALIDATION_FAILED",
                                    "UNAUTHENTICATED",
                                    "EMAIL_NOT_VERIFIED",
                                    "ACCOUNT_BLOCKED",
                                    "NOT_FOUND",
                                    "INVALID_TRANSITION",
                                    "CONFLICT",
                                    "BUSINESS_RULE_VIOLATION",
                                    "FEATURE_DISABLED",
                                    "RATE_LIMITED"
                                ]
                            },
                            "message": {
                                "type": "string"
                            },
                            "rule": {
                                "type": "string",
                                "description": "معرّف القاعدة في docs/04 عند BUSINESS_RULE_VIOLATION"
                            }
                        }
                    }
                }
            },
...
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ],
    "paths": {
...
        "/me/devices": {
            "post": {
                "operationId": "deviceStore",
                "tags": [
                    "المراجع"
                ],
                "summary": "deviceStore",
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AppMode"
                    },
                    {
                        "$ref": "#/components/parameters/IdempotencyKey"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "نجاح"
                    },
                    "400": {
                        "description": "خطأ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "deviceDestroy",
                "tags": [
                    "المراجع"
                ],
                "summary": "deviceDestroy",
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AppMode"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "نجاح"
                    },
                    "400": {
                        "description": "خطأ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}